+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- Fixed: Should not show "Offline support is disabled in development mode"
+  when offline isn't enabled
+  ([#448](https://github.com/MoOx/phenomic/issues/448))
+
 # 0.12.0 - 2016-05-05
 
 ## tl;dr;